moving standard controls used in the package item to the components directory
